+++ src/easydel/modules/phimoe/modeling_phimoe_flax.py
@@ -958,7 +958,7 @@ class FlaxPhiMoeForCausalLMModule(nn.Module):
 		else:
 			lm_logits = self.lm_head(outputs.last_hidden_state)
 
-		lm_logits = lm_logits.astype(jnp.float32)
+		
 
 		if not return_dict:
 			return (lm_logits,) + outputs[0:]
